Feature: New model FAQS

+++ src/controllers/ProductController.js
@@ -1,8 +1,9 @@
 const { PrismaClient } = require("@prisma/client");
 const prisma = new PrismaClient();
 
+
 const createProduct = async (req, res) => {
-  const { name, description, price, stock, sizes, color, gender } = req.body;
+  const { name, description, price, stock, sizes, color } = req.body;
 
   let images = [];
   let background = null;
@@ -29,7 +30,6 @@ const createProduct = async (req, res) => {
       images: JSON.stringify(images),
       background,
       color,
-      gender
     };
 
     let parsedSizes = [];
@@ -50,11 +50,13 @@ const createProduct = async (req, res) => {
       data: productData,
     });
 
+    // Crear tallas asociadas al producto y al género si se proporciona
     if (parsedSizes.length > 0) {
       const sizeData = parsedSizes.map((size) => ({
-        name: size.name,
+        sizeName: size.sizeName,
         stock: parseInt(size.stock),
         productId: product.id,
+        gender: size.gender || null, // Asignar el género de la talla si se proporciona
       }));
 
       await prisma.size.createMany({
@@ -69,23 +71,67 @@ const createProduct = async (req, res) => {
     });
   } catch (error) {
     console.error(error);
-    res.status(500).json({ error: "Fallo la creacion de la orden de compra" });
+    res.status(500).json({ error: "Fallo la creación del producto" });
   }
 };
+
+
+const getProductsByGenderAndSize = async (req, res) => {
+  const { gender } = req.query;
+
+  if (!gender) {
+    return res.status(400).json({
+      error: "El parámetro gender es requerido.",
+    });
+  }
+
+  try {
+    const products = await prisma.product.findMany({
+      where: {
+        sizes: {
+          some: {
+            gender: gender,
+          },
+        },
+      },
+      include: {
+        sizes: {
+          where: {
+            gender: gender,
+          },
+        },
+      },
+    });
+
+    if (products.length === 0) {
+      return res.status(404).json({
+        message: "No se encontraron productos para el género especificado.",
+      });
+    }
+
+    res.status(200).json(products);
+  } catch (error) {
+    console.error(error);
+    res.status(500).json({ error: "Error al consultar los productos" });
+  }
+};
+
 
 const getAllProducts = async (req, res) => {
   try {
     const products = await prisma.product.findMany({
       include: {
-        sizes: true,
+        sizes: true
       },
     });
+
     res.status(200).json(products);
   } catch (error) {
     console.error(error);
-    res.status(500).json({ error: "Fallo la consulta de los productos" });
+    res.status(500).json({ error: "Error al consultar los productos" });
   }
 };
+
 
 const getProductById = async (req, res) => {
   const { id } = req.params;
@@ -98,67 +144,80 @@ const getProductById = async (req, res) => {
         sizes: true,
       },
     });
-    if (product) {
-      res.status(200).json(product);
-    } else {
-      res.status(404).json({ error: "Producto no encontrado" });
+
+    if (!product) {
+      return res.status(404).json({ error: "Producto no encontrado" });
     }
+
+    res.status(200).json(product);
   } catch (error) {
     console.error(error);
-    res.status(500).json({ error: "No se encontro el producto" });
+    res.status(500).json({ error: "Fallo la consulta del producto" });
   }
 };
 
+
 const updateProduct = async (req, res) => {
   const { id } = req.params;
-  const { name, description, price, stock, sizes, gender } = req.body;
+  const { name, description, price, stock, sizes, color } = req.body;
 
   let images = [];
+  let background = null;
+
   if (req.files) {
-    images = req.files.map((file) => ({ url: file.filename }));
+    if (req.files.images) {
+      images = req.files.images.map((file) => ({ url: file.filename }));
+    }
+
+    if (req.files.background) {
+      background = req.files.background[0].filename;
+    }
   }
 
+  console.log(req.body);
+  console.log(images);
+  console.log(background);
+
   try {
+    // Obtener el producto actual para mantener las imágenes y el fondo si no se proporcionan nuevos datos
+    const currentProduct = await prisma.product.findUnique({
+      where: { id: parseInt(id) },
+    });
+
+    if (!currentProduct) {
+      return res.status(404).json({ error: "Producto no encontrado" });
+    }
+
     let productData = {
       name,
       description,
-      price: price ? parseFloat(price) : undefined,
-      images: images.length > 0 ? JSON.stringify(images) : undefined,
-      gender
+      price: parseFloat(price),
+      images: images.length > 0 ? JSON.stringify(images) : currentProduct.images,
+      background: background || currentProduct.background,
+      color,
     };
 
-    let parsedSizes = [];
-
-    if (sizes) {
-      try {
-        parsedSizes = JSON.parse(sizes);
-      } catch (error) {
-        return res.status(400).json({ error: "Tallas en formato incorrecto" });
-      }
+    if (stock) {
+      productData.stock = parseInt(stock);
     }
 
-    if (parsedSizes.length === 0) {
-      productData.stock = stock ? parseInt(stock) : undefined;
-    }
-
-    const product = await prisma.product.update({
-      where: {
-        id: parseInt(id),
-      },
+    const updatedProduct = await prisma.product.update({
+      where: { id: parseInt(id) },
       data: productData,
     });
 
-    if (parsedSizes.length > 0) {
+    if (sizes && sizes.length > 0) {
+      // Eliminar tallas existentes
       await prisma.size.deleteMany({
-        where: {
-          productId: product.id,
-        },
+        where: { productId: updatedProduct.id },
       });
 
-      const sizeData = parsedSizes.map((size) => ({
-        name: size.name,
+      // Crear nuevas tallas
+      const sizeData = sizes.map((size) => ({
+        sizeName: size.sizeName,
         stock: parseInt(size.stock),
-        productId: product.id,
+        productId: updatedProduct.id,
+        gender: size.gender || null,
       }));
 
       await prisma.size.createMany({
@@ -166,29 +225,51 @@ const updateProduct = async (req, res) => {
       });
     }
 
+    console.log(updatedProduct);
     res.status(200).json({
       message: "Producto actualizado exitosamente",
-      product,
+      updatedProduct,
     });
   } catch (error) {
     console.error(error);
-    res.status(500).json({ error: "Fallo la actualizacion del producto" });
+    res.status(500).json({ error: "Fallo la actualización del producto" });
   }
 };
 
 const deleteProduct = async (req, res) => {
   const { id } = req.params;
+
   try {
-    await prisma.product.delete({
-      where: {
-        id: parseInt(id),
-      },
+    const product = await prisma.product.findUnique({
+      where: { id: parseInt(id) },
+      include: { sizes: true, orders: true },
     });
+
+    if (!product) {
+      return res.status(404).json({ error: "Producto no encontrado" });
+    }
+
+    // Eliminar todas las tallas asociadas
+    await prisma.size.deleteMany({
+      where: { productId: parseInt(id) },
+    });
+
+    // Eliminar todas las asociaciones de OrderProduct
+    await prisma.orderProduct.deleteMany({
+      where: { productId: parseInt(id) },
+    });
+
+    // Finalmente, eliminar el producto
+    await prisma.product.delete({
+      where: { id: parseInt(id) },
+    });
+
     res.status(200).json({ message: "Producto eliminado exitosamente" });
   } catch (error) {
     console.error(error);
-    res.status(500).json({ error: "No se pudo eliminar el producto" });
+    res.status(500).json({ error: "Fallo la eliminación del producto" });
   }
 };
 
-module.exports = { createProduct, getAllProducts, getProductById, updateProduct, deleteProduct };
+
+module.exports = { createProduct, getProductsByGenderAndSize, getAllProducts, getProductById, updateProduct, deleteProduct };
